kubelet: pullmanager: move to a separate package
