+++ pkg/kubelet/images/pullmanager/fs_pullrecords.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package images
+package pullmanager
 
 import (
 	"bytes"
