+++ pkg/kubelet/images/pullmanager/image_pull_manager_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package images
+package pullmanager
 
 import (
 	"io/fs"
